To be is a to be the value of a variable

+++ 0x10-variadic_functions/3-print_all.c
@@ -28,7 +28,7 @@ void print_all(const char * const format, ...)
 					printf("%s%d", sep, va_arg(args, int));
 					break;
 				case 'f':
-					printf("%s%f", sep, va_arg(args, int));
+					printf("%s%f", sep, va_arg(args, double));
 					break;
 				case 's':
 					g = va_arg(args, char*);
